requirements.txt
Flask==3.0.3
Jinja2==3.1.4
SQLAlchemy==2.0.34
python-dotenv==1.0.1
pypdf==5.0.1
groq==0.11.0
requests==2.32.3
